+++ src/malt/malt.h
@@ -20,16 +20,16 @@ class Malt : public Meta {
    * @param ph pH of malt in purified water
    */
   // NOLINTNEXTLINE(*-magic-numbers)
-  explicit Malt(QString name = "", float mass = 0, float ebc = 0, float ph = 7);
+  explicit Malt(QString name = "", double mass = 0, double ebc = 0, double ph = 7);
   explicit Malt(const QJsonObject& json); /**< @brief Create Malt from JSON */
 
   // getter und setter
-  [[nodiscard]] float getMass() const; /**< @brief get malt amount in kg */
-  void setMass(float newMass);         /**< @brief set malt amount in kg */
-  [[nodiscard]] float getEbc() const;  /**< @brief get malt color in EBC */
-  void setEbc(float newEbc);           /**< @brief set malt color in EBC */
-  [[nodiscard]] float getPh() const;   /**< @brief get pH of malt in purified water */
-  void setPh(float newPh);             /**< @brief set pH of malt in purified water */
+  [[nodiscard]] double getMass() const; /**< @brief get malt amount in kg */
+  void setMass(double newMass);         /**< @brief set malt amount in kg */
+  [[nodiscard]] double getEbc() const;  /**< @brief get malt color in EBC */
+  void setEbc(double newEbc);           /**< @brief set malt color in EBC */
+  [[nodiscard]] double getPh() const;   /**< @brief get pH of malt in purified water */
+  void setPh(double newPh);             /**< @brief set pH of malt in purified water */
 
   enum type { upto200ebc, caramalz, roestmalz };
 
@@ -39,7 +39,7 @@ class Malt : public Meta {
    *
    * All calculations are based on https://maischemalzundmehr.de/index.php?inhaltmitte=exp_maischph
    */
-  static float calculatePh(float ebc, type type);
+  static double calculatePh(double ebc, type type);
 
   /**
    * @brief Calclates the pH in purified water based on ebc for malts up to 200 EBC
@@ -47,7 +47,7 @@ class Malt : public Meta {
    * x = -0,255 * ln(Malzfarbe [EBC]) + 6,156
    * \f]
    */
-  static float calcualtePhUpTo200Ebc(float ebc);
+  static double calcualtePhUpTo200Ebc(double ebc);
 
   /**
    * @brief Calclates the pH in purified water based on ebc for Carmalz
@@ -55,7 +55,7 @@ class Malt : public Meta {
    * x = -0,289 * ln(Malzfarbe [EBC]) + 6,069
    * \f]
    */
-  static float calcualtePhCaramalz(float ebc);
+  static double calcualtePhCaramalz(double ebc);
 
   /**
    * @brief Calclates the pH in purified water based on ebc for Röstamlz
@@ -63,16 +63,16 @@ class Malt : public Meta {
    * x = 4,65
    * \f]
    */
-  static float calcualtePhRoestmalz();
+  static double calcualtePhRoestmalz();
 
   // JSON conversion
   bool fromJson(const QJsonObject& json);   /**< @brief Update Malt from JSON */
   [[nodiscard]] QJsonObject toJson() const; /**< @brief convert this Malt to JSON */
 
  private:
-  float mass; /**< Amount of malt in kg */
-  float ebc;  /**< Color of malt in EBC */
-  float ph;   /**< pH of malt in purified water */
+  double mass; /**< Amount of malt in kg */
+  double ebc;  /**< Color of malt in EBC */
+  double ph;   /**< pH of malt in purified water */
 };
 
 #endif  // MALT_H
